load_nba_data add loaded without saving file to disk

+++ nba_on_court/nba_on_court.py
@@ -2,10 +2,12 @@ import warnings
 import re
 from pathlib import Path
 from itertools import product
-import urllib.request
+from urllib.request import urlopen
 import tarfile
 from functools import lru_cache
 from typing import Optional, List, Dict, Union, Sequence
+from io import BytesIO, TextIOWrapper
+import csv
 
 import numpy as np
 import pandas as pd
@@ -245,7 +247,9 @@ def load_nba_data(path: Union[Path, str] = Path.cwd(),
                                                 "shotdetail", "cdnnba", "nbastatsv3"),
                   seasontype: str = 'rg',
                   league: str = 'nba',
-                  untar: bool = False) -> None:
+                  untar: bool = False,
+                  in_memory: bool = False,
+                  use_pandas: bool = True) -> Optional[Union[List, pd.DataFrame]]:
     """
     Loading a nba play-by-play dataset from github repository https://github.com/shufinskiy/nba_data
 
@@ -256,9 +260,13 @@ def load_nba_data(path: Union[Path, str] = Path.cwd(),
         seasontype (str): Part of season: rg - Regular Season, po - Playoffs
         league (str): Name league: NBA or WNBA
         untar (bool): Logical: do need to untar loaded archive
+        in_memory (bool): Logical: If True dataset is loaded into workflow, without saving file to disk
+        use_pandas (bool): Logical: If True dataset is loaded how pd.DataFrame, else List[List[str]]. Ignore if in_memory=False
 
     Returns:
-        None
+        Optional[pd.DataFrame, List]: If in_memory=True and use_pandas=True return dataset how pd.DataFrame.
+        If use_pandas=False return dataset how List[List[str]]
+        If in_memory=False return None
     """
     if isinstance(path, str):
         path = Path(path)
@@ -266,6 +274,9 @@ def load_nba_data(path: Union[Path, str] = Path.cwd(),
         seasons = (seasons,)
     if isinstance(data, str):
         data = (data,)
+
+    if (len(data) > 1) & in_memory:
+        raise ValueError("Parameter in_memory=True available only when loading a single data type")
 
     if seasontype == 'rg':
         need_data = tuple(["_".join([data, str(season)]) for (data, season) in product(data, seasons)])
@@ -285,7 +296,7 @@ def load_nba_data(path: Union[Path, str] = Path.cwd(),
 
     need_data = [file for (file, not_exist) in zip(need_data, not_exists) if not_exist]
 
-    with urllib.request.urlopen("https://raw.githubusercontent.com/shufinskiy/nba_data/main/list_data.txt") as f:
+    with urlopen("https://raw.githubusercontent.com/shufinskiy/nba_data/main/list_data.txt") as f:
         v = f.read().decode('utf-8').strip()
 
     name_v = [string.split("=")[0] for string in v.split("\n")]
@@ -294,15 +305,38 @@ def load_nba_data(path: Union[Path, str] = Path.cwd(),
     need_name = [name for name in name_v if name in need_data]
     need_element = [element for (name, element) in zip(name_v, element_v) if name in need_data]
 
+    if in_memory:
+        if use_pandas:
+            table = pd.DataFrame()
+        else:
+            table = []
     for i in range(len(need_name)):
-        t = urllib.request.urlopen(need_element[i])
-        with path.joinpath("".join([need_name[i], ".tar.xz"])).open(mode='wb') as f:
-            f.write(t.read())
-        if untar:
-            with tarfile.open(path.joinpath("".join([need_name[i], ".tar.xz"]))) as f:
-                f.extract("".join([need_name[i], ".csv"]), path)
+        with urlopen(need_element[i]) as response:
+            if response.status != 200:
+                raise Exception(f"Failed to download file: {response.status}")
+            file_content = response.read()
+            if in_memory:
+                with tarfile.open(fileobj=BytesIO(file_content), mode='r:xz') as tar:
+                    csv_file_name = "".join([need_name[i], ".csv"])
+                    csv_file = tar.extractfile(csv_file_name)
+                    if use_pandas:
+                        table = pd.concat([table, pd.read_csv(csv_file)], axis=0, ignore_index=True)
+                    else:
+                        csv_reader = csv.reader(TextIOWrapper(csv_file, encoding="utf-8"))
+                        for row in csv_reader:
+                            table.append(row)
+            else:
+                with path.joinpath("".join([need_name[i], ".tar.xz"])).open(mode='wb') as f:
+                    f.write(file_content)
+                if untar:
+                    with tarfile.open(path.joinpath("".join([need_name[i], ".tar.xz"]))) as f:
+                        f.extract("".join([need_name[i], ".csv"]), path)
 
-            path.joinpath("".join([need_name[i], ".tar.xz"])).unlink()
+                    path.joinpath("".join([need_name[i], ".tar.xz"])).unlink()
+    if in_memory:
+        return table
+    else:
+        return None
 
 
 def _concat_description(homedescription: pd.Series,
